Add files via upload
This is the GRAND-FINALE report for the IBM Data Science Capstone project. It was originally PPTX with animated GIFS for the interactive maps (showcasing what they can do) but for font compatibility reasons I turned this into a PDF document and compressed it because it was still too big for Git. If you need the original let me know.