Add files via upload
This is a PDF of the PPTX I created for the project. The PPTX has animated GIFs that showcase the interactivity of the generated maps (zoom or selection). I was concerned about the transferability of the fonts. The converted PDF was too large for GitHub so I compressed it. If you need access to the original please contact me.